gfx/layers/apz/test/mochitest/test_group_touchevents-4.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Various touch tests that spawn in new windows (4)</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="apz_test_native_event_utils.js"></script>
  <script type="application/javascript" src="apz_test_utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

var touch_action_prefs = getPrefs("TOUCH_ACTION");

var subtests = [
  // clicking on element with :active::after CSS property
  {"file": "helper_bug1473108.html"},
  // Resetting isFirstPaint shouldn't clobber the visual viewport
  {"file": "helper_bug1509575.html", "prefs": [["apz.allow_zooming", true],
                                               ...getPrefs("TOUCH_EVENTS:PAN")]},
  // Exercise one of the main-thread touch-action determination codepaths.
  {"file": "helper_bug1506497_touch_action_fixed_on_fixed.html", "prefs": touch_action_prefs},
  // Add new subtests here. If this starts timing out because it's taking too
  // long, create a test_group_touchevents-5.html file. Refer to 1423011#c57
  // for more details.
];

if (isApzEnabled()) {
  ok(window.TouchEvent, "Check if TouchEvent is supported (it should be, the test harness forces it on everywhere)");
  if (getPlatform() == "android") {
    // This has a lot of subtests, and Android emulators are slow.
    SimpleTest.requestLongerTimeout(2);
  }

  SimpleTest.waitForExplicitFinish();
  window.onload = function() {
    runSubtestsSeriallyInFreshWindows(subtests)
    .finally(SimpleTest.finish);
  };
}

  </script>
</head>
<body>
</body>
</html>
